send emails when inviting people
